+++ src/handovergap/stores/__init__.py
@@ -1,3 +1,3 @@
-from handovergap.stores.tidb import TiDBStore
+from handovergap.stores.tidb import TiDBSchemaState, TiDBStore, TiDBStoreOperationError
 
-__all__ = ["TiDBStore"]
+__all__ = ["TiDBSchemaState", "TiDBStore", "TiDBStoreOperationError"]